% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGP.R
\name{generate_data}
\alias{generate_data}
\title{data generation process given some information using R}
\usage{
generate_data(n, p, T, r, h, epsilon_bar)
}
\arguments{
\item{n}{sample size of our task}

\item{p}{sample dimension of our task}

\item{T}{task size for our MTL}

\item{r}{intrinsic dimension for linear coefficients}

\item{h}{similarity concentration level for representation matrix A}

\item{epsilon_bar}{projection quantile level for Spectral 
Method(eg:0.1 represent projection to 0.9-quantile level)}
}
\value{
data contains sample and some parameter information
}
\description{
data generation process given some information using R
}
\examples{
\dontrun{
    data <- generate_data(n, p, T, r, h, epsilon_bar)
    beta_true <- data$beta  # True beta values from the model
    beta_final_spectral <- Spectral_Method_kr(data$X, data$Y, 
    T, r, epsilon_bar, gamma = 1)
}
}
