% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbs_sampler}
\alias{gibbs_sampler}
\title{Gibbs sampling algorithm for sampling from the joint distribution of 
two variables x and y.}
\usage{
gibbs_sampler(n, a, b, num_samples, burn_in)
}
\arguments{
\item{n}{Integer. The total number of trials for the binomial distribution in each step.}

\item{a}{Double. The first shape parameter for the Beta distribution.}

\item{b}{Double. The second shape parameter for the Beta distribution.}

\item{num_samples}{Integer. The number of posterior samples to draw after burn-in.}

\item{burn_in}{Integer. The number of initial iterations to discard as burn-in.}
}
\value{
NumericMatrix A matrix of size (num_samples, 2), where each row contains
the sampled values of x and y.
}
\description{
The function performs sampling from the conditionaldistributions p(x | y) and p(y | x) iteratively.
}
\examples{
\dontrun{
samples <- gibbs_sampler(10, 2, 2, 1000, 200)
}
}
