% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{projection}
\alias{projection}
\title{Projection onto the ℓ2-ball}
\usage{
projection(T, beta_hat, X, epsilon_bar)
}
\arguments{
\item{T}{An integer representing the number of iterations or the length of `beta_hat`.}

\item{beta_hat}{A list of vectors (length T), where each element is a vector of coefficients.}

\item{X}{A matrix of input data, where rows represent observations and columns represent features.}

\item{epsilon_bar}{A numeric value (between 0 and 1) used to determine the quantile value 
for projection based on the upper quantile of the norms.}
}
\value{
A matrix of projected coefficients, with each column representing the projected 
        vector corresponding to each element in `beta_hat`.
}
\description{
This function performs a projection of a sequence of vectors (betas) onto the ℓ2-ball 
based on the upper quantile of the norms of the vectors. The projection scales vectors
with a norm greater than the quantile value to fit within the ℓ2-ball of radius defined 
by the quantile value.
}
\examples{
{
\dontrun{
# Example usage:
projection_result <- projection(100, beta_hat_list, X_matrix, 0.05)
# Visualizing the result
plot(projection_result[, 1], type = 'l')
plot(projection_result[, 2], type = 'l')
}
}
}
