% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_r.R
\name{r_estimation}
\alias{r_estimation}
\title{An estimation of intrinsic dimension of linear coefficients}
\usage{
r_estimation(Y, X, T, epsilon_bar, T1 = 0.5, T2 = 0.25)
}
\arguments{
\item{Y}{Response variable with dimension n(sample size) and T(task size)}

\item{X}{Explation variable array with dimension n(sample size),
p(dimension),T(task size)}

\item{T}{task size for MTL}

\item{epsilon_bar}{projection quantile level for Spectral 
Method(eg:0.1 represent projection to 0.9-quantile level)}

\item{T1}{thresholding penalty parameter for sqrt((p+logT)/n),default 0.5}

\item{T2}{thresholding penalty parameter for R*epsilon_bar(default 0.25)}
}
\value{
intrinsic dimension estimation for linear coefficient
}
\description{
An estimation of intrinsic dimension of linear coefficients
}
\examples{
\dontrun{
    X <- array(runif(5 * 100 * 10), dim = c(100, 10, 5))
    Y <- array(runif(5 * 100), dim = c(100, 5))
    r_hat <- r_estimation(Y,X,5,0.1)
    print(r_hat)
}
}
