% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_task.R
\name{single_task}
\alias{single_task}
\title{Single Task Linear Regression}
\usage{
single_task(Y, X, T)
}
\arguments{
\item{Y}{A matrix of response variables of size \code{n} x \code{T}, where each column represents 
the responses for a particular task.}

\item{X}{A 3D array of input features of size \code{n} x \code{p} x \code{T}, where each slice corresponds 
to the feature matrix for a particular task.}

\item{T}{The number of tasks. This determines how many times the regression will be performed.}
}
\value{
A list containing the estimated coefficients for each task
}
\description{
This function performs linear regression for each task
 in a multi-task setting.
}
\examples{
\dontrun{
    # Example usage:
    beta_hat <- single_task(Y = data$Y, X = data$X, T = 10)
    # beta_hat contains the estimated coefficients for each task
}

}
