% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral_Method_kr.R
\name{Spectral_Method_kr}
\alias{Spectral_Method_kr}
\title{Spectral Method for Representation Multitask Learning(RMTL) 
when know intrinsic dimension r}
\usage{
Spectral_Method_kr(X, Y, T, r, epsilon_bar, gamma)
}
\arguments{
\item{X}{Explation variable array with dimension 
n(sample size),p(dimension),T(task size)}

\item{Y}{Response variable with dimension 
n(sample size) and T(task size)}

\item{T}{task size for MTL}

\item{r}{intrisic dimension for linear coefficient}

\item{epsilon_bar}{projection quantile level for Spectral Method}

\item{gamma}{regularization parameter for bias regularization step}
}
\value{
linear coefficients for T tasks using RMTL method
}
\description{
Spectral Method for Representation Multitask Learning
 when know intrinsic dimension r
}
\examples{
\dontrun{
    X <- array(runif(5 * 100 * 10), dim = c(100, 10, 5))
    Y <- array(runif(5 * 100), dim = c(100, 5))
    beta_final <- Spectral_Method_kr(X,Y,5,3,0.1,1)
    print(beta_final)
}
}
