% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectral_Method_ur.R
\name{Spectral_Method_ur}
\alias{Spectral_Method_ur}
\title{Spectral Method for Representation Multitask Learning(RMTL) 
when unknow intrinsic dimension r}
\usage{
Spectral_Method_ur(X, Y, T, epsilon_bar, gamma, T1 = 0.5, T2 = 0.25)
}
\arguments{
\item{X}{Explation variable array with dimension 
n(sample size),p(dimension),T(task size)}

\item{Y}{Response variable with dimension 
n(sample size) and T(task size)}

\item{T}{task size for MTL}

\item{epsilon_bar}{projection quantile level for Spectral Method}

\item{gamma}{regularization parameter for bias regularization step}

\item{T1}{thresholding penalty parameter for sqrt((p+logT)/n),default 0.5}

\item{T2}{thresholding penalty parameter for R*epsilon_bar(default 0.25)}
}
\value{
linear coefficients for T tasks using RMTL method
}
\description{
Spectral Method for Representation Multitask Learning
 when unknow intrinsic dimension r
}
\examples{
\dontrun{
    X <- array(runif(5 * 100 * 10), dim = c(100, 10, 5))
    Y <- array(runif(5 * 100), dim = c(100, 5))
    beta_final <- Spectral_Method_ur(X,Y,5,0.1,1)
    print(beta_final)
}
}
