% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVD.R
\name{SVD}
\alias{SVD}
\title{Singular Value Decomposition (SVD) for Matrix Decomposition}
\usage{
SVD(B_hat, r)
}
\arguments{
\item{B_hat}{A matrix of size \code{n} x \code{p} to undergo singular value decomposition.}

\item{r}{The intrinsic dimension for the decomposition, i.e., the number of left singular vectors to extract.}
}
\value{
A matrix \code{A_hat} containing the first \code{r} columns of the left singular vector matrix \code{U}
        from the SVD of \code{B_hat}.
}
\description{
This function performs Singular Value Decomposition (SVD) on a given matrix \code{B_hat}.
It then extracts the first \code{r} columns of the left singular vector matrix \code{U}, 
which corresponds to the intrinsic dimension of the matrix.
}
\examples{
\dontrun{
    # Example usage:
    A_hat <- SVD(B_hat = some_matrix, r = 5)
    # A_hat will contain the first 5 left singular vectors of B_hat
}

}
