% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransRL_Spectral.R
\name{TransRL_Spectral}
\alias{TransRL_Spectral}
\title{Spectral Method for Representation Multitask Learning(RMTL) 
applies to new transfer learning task}
\usage{
TransRL_Spectral(
  X_source,
  Y_source,
  X_target,
  Y_target,
  T,
  epsilon_bar,
  gamma,
  T1 = 0.5,
  T2 = 0.25
)
}
\arguments{
\item{X_source}{Explation source variable array with dimension 
n(sample size),p(dimension),T(task size)}

\item{Y_source}{Response source variable with dimension n(sample size) 
and T(task size)}

\item{X_target}{New explation variable with dimension n0(sample size) and p(variable dimension)}

\item{Y_target}{New response variable with dimension n0}

\item{T}{task size for MTL}

\item{epsilon_bar}{projection quantile level for Spectral Method}

\item{gamma}{regularization parameter for bias regularization step}

\item{T1}{thresholding penalty parameter for sqrt((p+logT)/n),default 0.5}

\item{T2}{thresholding penalty parameter for R*epsilon_bar(default 0.25)}
}
\value{
linear coefficients for new task
}
\description{
Spectral Method for Representation Multitask Learning
 applies to new transfer learning task
}
\examples{
\dontrun{
    X_source <- array(runif(5 * 100 * 10), dim = c(100, 10, 5))
    Y_source <- array(runif(5 * 100), dim = c(100, 5))
    Y_target  <- runif(n)
    X_target <- matrix(runif(n*p),ncol = p,nrow = n)
    beta0 <- TransRL_Spectral(X_source,Y_source,X_target,Y_target,T,epsilon_bar,
    gamma,T1=0.5,T2=0.25)
    print(beta0)
}
}
